requirements.txt
fastapi
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.7.4
celery==5.4.0
redis==5.0.8
trafilatura==1.12.2
google-generativeai==0.8.3
